+++ skforecast/ForecasterAutoreg/tests/test_set_fit_kwargs.py
@@ -1,0 +1,19 @@
+# Unit test set_fit_kwargs ForecasterAutoreg
+# ==============================================================================
+from skforecast.ForecasterAutoreg import ForecasterAutoreg
+from lightgbm import LGBMRegressor
+
+def test_set_fit_kwargs():
+    """
+    Test set_fit_kwargs method.
+    """
+    forecaster = ForecasterAutoreg(
+                    regressor = LGBMRegressor(),
+                    lags = 3,
+                    fit_kwargs = {'categorical_feature': 'auto'}
+                 )
+    new_fit_kwargs = {'categorical_feature': ['exog']}
+    forecaster.set_fit_kwargs(new_fit_kwargs)
+    expected = {'categorical_feature': ['exog']}
+    results = forecaster.fit_kwargs
+    assert results == expected